requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.7
django-storages==1.14.2
boto3==1.29.7
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
whitenoise==6.6.0
gunicorn==21.2.0
